+++ StrideLiveEditor.Avalonia/DataTypeEditors/Vector2Editor.xaml.cs
@@ -38,9 +38,9 @@ namespace StrideLiveEditor.Avalonia.DataTypeEditors
             var value = (Vector2)ComponentProperty.GetValue(Component);
 
             if ((!editorWindowIsActive || !X.IsFocused) && GetFloat(X.Value) != value.X)
-                X.Value = value.X;
+                X.Value = (decimal)value.X;
             if ((!editorWindowIsActive || !Y.IsFocused) && GetFloat(Y.Value) != value.Y)
-                Y.Value = value.Y;
+                Y.Value = (decimal)value.Y;
         }
 
         private void InitializeComponent()
